requirements.txt
flask
flask-login
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf
Werkzeug
python-dateutil
